This adds benchmarks for comparing standard RPC to RPC-over-Keyless. If you checkout https://github.com/cloudflare/gokeyless/commit/255c0f3eb63bc54b7fb6b48798c3aed6da1bbf80 and run `cd tests && go test -bench .`, you'll see that RPC-over-Keyless is up to 10x as slow as RPC:
```
pkg: github.com/cloudflare/gokeyless/tests
BenchmarkKeylessDummyRPC-4   	   10000	    169289 ns/op	   32688 B/op	     471 allocs/op
BenchmarkDummyRPC-4          	  100000	     16991 ns/op	     400 B/op	      12 allocs/op
```
Commit https://github.com/cloudflare/gokeyless/pull/245/commits/6efd7c35b7207a435a7bb2eac6c001ce7a7f999c makes RPC-over-Keyless faster by making encoding and decoding of RPCs stateful. This improves performance by up to a factor of 2:
```
pkg: github.com/cloudflare/gokeyless/tests
BenchmarkKeylessDummyRPC-4   	   20000	     81775 ns/op	   10935 B/op	      48 allocs/op
BenchmarkDummyRPC-4          	  100000	     16408 ns/op	     388 B/op	      12 allocs/op
```
I think there's definitely room for improvement, but it would take some work. This PR is meant to be minimally invasive.

NOTE: this PR changes the wire protocol. If an out-of-date client tries to use RPC with a current server (or vice versa), then the call will fail.