@cbroglie 2018-07-29
Changing the wire protocol is unacceptable if it can't be done in a backwards compatible way.