@cjpatton 2018-07-30
That's what I figured. One way we could maintain backwards compatibility is to add a new opcode for doing stateful RPC.

But is this really worth doing? Given the current use cases for RPC-over-Keyless, perhaps not. But if this code path is latency sensitive, then to me it seems the current performance is cause for concern.