@cjpatton 2018-07-30
For signing DCs using RPC-over-Keyless, a similar benchmark clocks in at 245939 ns/op. Is that Ok, do you think? (Using this speed up gets it down to 155192 ns/op.)