@cjpatton 2018-07-30
Actually, I think it's worth at least merging the benchmark: **https://github.com/cloudflare/gokeyless/pull/245/commits/255c0f3eb63bc54b7fb6b48798c3aed6da1bbf80**