@cbroglie 2019-05-20
This changes the wire protocol so it cannot be merged. The benchmark included here proves there is no need for optimizations.